+++ backend/src/test/kotlin/org/climatechangemakers/act/common/util/WithRetryTest.kt
@@ -3,6 +3,8 @@ package org.climatechangemakers.act.common.util
 import okhttp3.MediaType
 import okhttp3.ResponseBody
 import org.climatechangemakers.act.feature.findlegislator.util.suspendTest
+import org.climatechangemakers.act.feature.membership.model.AirtableRecord
+import org.climatechangemakers.act.feature.membership.model.AirtableResponse
 import retrofit2.HttpException
 import retrofit2.Response
 import kotlin.test.Test
@@ -10,10 +12,6 @@ import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 
 class WithRetryTest {
-
-  @Test fun `withRetry exponentially backs off`() {
-    // TODO(kcianfarini) write test with delaycontroller
-  }
 
   @Test fun `withRetry throws exception for most recent failure`() = suspendTest {
     val throwable = assertFailsWith<HttpException> {
@@ -40,5 +38,17 @@ class WithRetryTest {
     }
 
     assertEquals(400, throwable.code())
+  }
+
+  @Test fun `withRetry returns first successful response`() = suspendTest {
+    val response = withRetry(2) { attempt ->
+      if (attempt == 0) {
+        Response.error(429, ResponseBody.create(MediaType.get("application/json"), "please retry"))
+      } else {
+        Response.success(AirtableResponse(records = listOf(AirtableRecord(id = "foo"))))
+      }
+    }
+
+    assertEquals("foo", response.records.first().id)
   }
 }
