+++ backend/src/test/kotlin/org/climatechangemakers/act/feature/findlegislator/util/Functions.kt
@@ -2,6 +2,8 @@ package org.climatechangemakers.act.feature.findlegislator.util
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.runBlocking
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
 
 fun suspendTest(block: suspend CoroutineScope.() -> Unit) {
   runBlocking { block() }
