+++ backend/src/test/kotlin/org/climatechangemakers/act/feature/membership/manager/AirtableMembershipManagerTest.kt
@@ -2,15 +2,15 @@ package org.climatechangemakers.act.feature.membership.manager
 
 import okhttp3.MediaType
 import okhttp3.ResponseBody
-import org.climatechangemakers.act.common.util.withRetry
 import org.climatechangemakers.act.feature.action.manager.FakeActionTrackerManager
 import org.climatechangemakers.act.feature.findlegislator.util.suspendTest
 import org.climatechangemakers.act.feature.membership.model.AirtableRecord
 import org.climatechangemakers.act.feature.membership.model.AirtableResponse
+import org.climatechangemakers.act.feature.membership.service.AirtableFormula
+import org.climatechangemakers.act.feature.membership.service.AirtableService
 import org.climatechangemakers.act.feature.membership.service.FakeAirtableService
 import org.junit.Test
 import retrofit2.Response
-import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
@@ -29,14 +29,23 @@ class AirtableMembershipManagerTest {
   }
 
   @Test fun `checkMembership retries`() = suspendTest {
-    val response = withRetry(2) { attempt ->
-      if (attempt == 0) {
-        Response.error(429, ResponseBody.create(MediaType.get("application/json"), "please retry"))
-      } else {
-        Response.success(AirtableResponse(records = listOf(AirtableRecord(id = "foo"))))
+    val fakeService = object : AirtableService {
+
+      private var hasFailed = false
+
+      override suspend fun checkMembership(formula: AirtableFormula.FilterByEmailFormula): Response<AirtableResponse> {
+        val result = if (hasFailed) {
+          Response.success(AirtableResponse(listOf(AirtableRecord(formula.email))))
+        } else {
+          Response.error(429, ResponseBody.create(MediaType.get("application/json"), "Failed"))
+        }
+
+        hasFailed = !hasFailed
+
+        return result
       }
     }
-
-    assertEquals("foo", response.records.first().id)
+    val manager = AirtableMembershipManager(FakeActionTrackerManager(), fakeService)
+    assertTrue(manager.checkMembership("foo@bar.com"))
   }
 }
